[profiler][small] Handle empty trace (#58013)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58013

Add a test case and a fix (legacy profiler) for empty trace handling

Test Plan: python test/test_profiler.py

Reviewed By: gdankel

Differential Revision: D28345388

Pulled By: ilia-cher

fbshipit-source-id: 4727589ab83367ac8b506cc0f186e5292d974671

+++ torch/autograd/profiler.py
@@ -247,10 +247,10 @@ class EventList(list):
                                                evt.thread, next_id))
                     # Note: use torch.profiler to get device kernel trace
                     next_id += 1
-
-            # remove trailing whitespace and comma
-            f.seek(f.tell() - 2, os.SEEK_SET)
-            f.truncate()
+            if len(self) > 0:
+                # remove trailing whitespace and comma
+                f.seek(f.tell() - 2, os.SEEK_SET)
+                f.truncate()
             f.write("]")
 
     def supported_export_stacks_metrics(self):
